resume_checker/serializers.py
from rest_framework import serializers
from resume_checker.models import Vacancy, Resume


class VacancySerializer(serializers.ModelSerializer):
    class Meta:
        model = Vacancy
        fields = '__all__'


class ResumeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Resume
        fields = '__all__'
