requirements.txt
Django==4.2.8
djangorestframework==3.14.0
langchain==0.0.352
langchain_core==0.1.3
openai==1.6.0
pydantic==2.5.2
python-dotenv==1.0.0
